refactored cart-view component
