+++ client/components/index.js
@@ -6,7 +6,7 @@
 
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
-export {default as Cart} from './cart'
-export {default as CartItems} from './cart-items'
-export {default as Item} from './item'
+export {default as Cart} from './orders/cart'
+export {default as CartView} from './orders/cart-view'
+export {default as Item} from './orders/item'
 export {Login, Signup} from './auth-form'
