+++ client/store/index.js
@@ -14,3 +14,5 @@ const store = createStore(reducer, middleware)
 
 export default store
 export * from './user'
+export * from './order'
+export * from './product'
